backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.6
constantly==15.1.0
cryptography==41.0.1
cssselect2==0.7.0
daphne==4.0.0
defusedxml==0.7.1
Django==4.2.2
django-cors-headers==4.1.0
django-environ==0.10.0
django-extensions==3.2.3
django-sslserver==0.22
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fonttools==4.41.1
google-auth==2.21.0
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
infobip-api-python-sdk==5.0.1
joblib==1.3.1
lxml==4.9.3
msgpack==1.0.5
numpy==1.25.0
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.1
Pillow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pydotplus==2.0.2
pydyf==0.7.0
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
PyJWT==2.7.0
PyMySQL==1.0.3
pyOpenSSL==23.2.0
pyparsing==3.1.0
pypdf==3.14.0
pyphen==0.14.0
pypng==0.20220715.0
pyproj==3.6.0
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
redis==4.6.0
regex==2022.3.15
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
service-identity==23.1.0
shapely==2.0.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
svglib==1.5.1
threadpoolctl==3.2.0
tinycss2==1.2.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==1.26.16
webencodings==0.5.1
xhtml2pdf==0.2.11
zope.interface==6.0
zopfli==0.2.2
